Add a hook that allows us to force a protocol to be expressed as the protocol number instead of the protocol name. This helps with cases in which a protocol name is mapped to a different protocol number due to platform differences (e.g., ipip is protocol 94 in Linux due to /etc/protocols, but the network industry identifies ipip as protocol number 4).
This bug refers to iptables, so we could insert a modification on the
iptables Capirca generator. However, it seems we can add this to be base class
if we have similar problems in the future with other platforms.

In addition, remove protocol map ipencap which is unused and confusing. Stick
to ipip for IP encapsulation and use protocol number 4 for it.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=130657973