(retry of https://github.com/google/capirca/pull/83)
Summary:
This code adds token and sub token validation for all platforms.  I made a good effort to inspect the code and capture all of the token use cases, but there may be some omissions. Essentially what this will do is define a strict set of tokens that each platform implements, instead of relying on the platform to maybe provide what it supports.  Making it mandatory that a platform defines what it implements will make testing easier and also create a better experience for users writing rules against a platform.  They will know immediately that the term they wrote will actually render as they intended it to.

These tokens are inspected in __init__ of AclGenerator before the the respective platform _TranslatePolicy.  This will prevent unexpected renders of policy, as sometimes platforms can silently drop tokens the user provides. It also enforces actions, in a potentially disagreeable way, but does so explicitly.  Users will know straight away if a token or sub token they are using will not work as expected, and explicitly as they have defined it in the term itself.